Ensure raw_email file representations are only deleted if they exist

+++ app/models/raw_email.rb
@@ -63,7 +63,7 @@ class RawEmail < ActiveRecord::Base
   end
 
   def destroy_file_representation!
-    File.delete(filepath)
+    File.delete(filepath) if File.exists?(filepath)
   end
 
   private
